refactor: extract UI from search_command.py into menu classes
- Create dedicated menu classes for search-related UI:
  - SearchParametersMenu
  - SearchResultsMenu
  - SearchMenu (main menu)
- Remove UI logic from search_command.py
- Keep only business logic in command file
